master: Weight API Functionality completed

+++ src/main/java/com/company/airport/models/weightapi/WeightResponse.java
@@ -3,14 +3,13 @@ package com.company.airport.models.weightapi;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.ToString;
 
 @Data
 @AllArgsConstructor
 @Builder
 public class WeightResponse {
-    public Long cargoWeight;
-    public Long baggageWeight;
-    public Long totalWeight;
+    public Double cargoWeight;
+    public Double baggageWeight;
+    public Double totalWeight;
     public String error;
 }
